Refactor tests for clarity and extract Foo class
This commit performs a structural refactor of the StaticMockingExperimentTest to enhance code readability and maintainability. By extracting the Foo class into its own file, we follow the Single Responsibility Principle, making the codebase more organized. The Foo class, previously nested within a test class, now resides in its own dedicated file, separating the business logic from the testing logic.

This change not only improves the organization of the code but also makes it easier to navigate and understand. The StaticMockingExperimentTest has been updated to reflect this extraction, focusing solely on testing functionalities, including experimental approaches to mocking static methods with Mockito.

Key changes include:
- Extraction of the Foo class into its own file.
- Update of StaticMockingExperimentTest to remove the nested Foo class and adjust to the new structure.
- Preservation of all test functionalities and the introduction of clearer separation between business logic and test logic.

This refactor demonstrates the project's commitment to clean code practices and enhances the ability for future extensions and maintenance.
